video_shot_segmenter/requirements.txt
PySceneDetect
numpy
opencv-python # Often a dependency or useful with PySceneDetect 